requirements.txt
alpaca-py>=0.19.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pathlib
matplotlib
tqdm
pytz

# Core
python-dotenv>=1.0.0
pytz>=2023.3

# Alpaca
alpaca-py>=0.19.0

# Numerical / indicators
numpy>=1.24.0

# (Optional but recommended later)
pandas>=2.0.0
matplotlib>=3.8.0
